requirements.txt
flask==2.3.3
flask-cors==4.0.0
google-cloud-translate==3.11.1
google-api-core==2.11.0
pyttsx3==2.90
SpeechRecognition==3.10.0
pyaudio==0.2.13
python-dotenv==1.0.0
requests==2.31.0
werkzeug==2.3.7
gunicorn==21.2.0
